MapDataFetcher/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-login==0.6.2
flask-mail==0.9.1
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
gunicorn==21.2.0
matplotlib==3.8.0
numpy==1.26.0
pandas==2.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
reportlab==4.0.4
requests==2.31.0
scikit-learn==1.3.0
trafilatura==1.6.1 